refactor: validator 디렉토리 위치 수정
